app/requirements.txt
flask
livereload
python-dotenv
flask-wtf